프록시 구축
